open player by route
